cfe__15__5_46__JSON WEB Token Authentication with/backend/requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.0.10
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
